fix: resolve Xray config duplicate tag error and compilation issues
- Fix duplicate 'api-in' tag error by removing existing inbounds before adding
- Fix C compilation errors in hev-socks5-tunnel (comment out non-existent timeout functions)
- Replace batch UDP functions with single-message functions in hev-socks5-session-udp.c
- Fix Kotlin compilation errors (missing imports, type mismatches, StateFlow usage)
- Add missing extension function imports in PerformanceMetricsRepository
- Fix type conversions in AdaptivePerformanceTuner
- Add missing Alignment import in CustomProfileEditScreen
- Fix StateFlow usage in PerformanceViewModel

+++ app/src/main/kotlin/com/simplexray/an/performance/optimizer/AdaptivePerformanceTuner.kt
@@ -125,7 +125,7 @@ class AdaptivePerformanceTuner(
                 return ProfileRecommendation(
                     profile = PerformanceProfile.Gaming,
                     reason = "High latency detected (${metrics.latency}ms). Gaming profile optimizes for low latency.",
-                    confidence = calculateConfidence(metrics.latency, highLatencyThreshold, 500),
+                    confidence = calculateConfidence(metrics.latency.toFloat(), highLatencyThreshold.toFloat(), 500f),
                     trigger = RecommendationTrigger.HighLatency(metrics.latency)
                 )
             }
